requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.1.4
django-crispy-forms==1.14.0
djangorestframework==3.14.0
Pillow==9.3.0
psycopg==3.1.7
psycopg-binary==3.1.6
psycopg2-binary==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
typing-extensions==4.4.0
